backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Diego Sesmas <diegos7167@gmail.com>"]
maintainers = ["Diego Sesmas <diegos7167@gmail.com>"]
repository = "https://github.com/your-repo-url"  # Update this with your repository URL
license = "MIT"  # Update this if you have a different license
readme = "README.md"  # Update this if you have a different readme file
keywords = []  # Add keywords if needed
packages = [{ include = "app", from = "." }]

[tool.poetry.scripts]
# Add any entry points if needed

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.103.1"
uvicorn = { extras = ["standard"], version = "^0.23.2" }
sqlalchemy = "^2.0.20"
asyncpg = "^0.28.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.3"
argon2-cffi = "^23.1.0"
pydantic-extra-types = "^2.1.0"
httpx-oauth = "^0.13.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
redfin = "^0.1.1"
pandas = "^2.1.1"
scikit-learn = "1.3.0"
google-auth-httplib2 = "^0.1.1"
google-api-python-client = "^2.105.0"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[tool.ruff]
line-length = 120  # Update this if you have a different line length

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
